camera_trigger: Use proper types

+++ src/drivers/camera_trigger/interfaces/src/camera_interface.cpp
@@ -44,7 +44,7 @@ void CameraInterface::get_pins()
 		return;
 	}
 
-	int pin_list;
+	long pin_list;
 	param_get(_p_pin, &pin_list);
 
 	// Set all pins as invalid
